Use 'DevKitGameID' in shortcuts.vdf as a container for the GOG game ID
Use this as an alternative method for identifing existing GOG games in
the shortcuts file that is not tied to the path or game name. Also this
value is not used an always empty, so this is a way to make use of this
empty variable. Any other non-standard keys added to the shortcuts file
are pruned by steam on app close!